requirements.txt
beautifulsoup4==4.9.3
Brotli==1.0.9
bs4==0.0.1
certifi==2020.12.5
click==7.1.2
dash==1.17.0
dash-core-components==1.13.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
elasticsearch==7.10.1
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
libsass==0.20.1
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.1.4
plotly==4.12.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
retrying==1.3.3
scss==0.8.73
six==1.15.0
soupsieve==2.0.1
urllib3==1.26.2
Werkzeug==1.0.1
